+++ resources/public/index.html
@@ -102,8 +102,8 @@
 <div class="ui" id="ui">
     <div>
         <div class="login-overlay">
-            You are not signed in. Sign in with <a href="/signin/reddit">Reddit</a> or <a
-                href="/signin/google">Google</a>.
+            You are not signed in. Sign in with <a href="/signin/reddit">Reddit</a> | <a
+                href="/signin/google">Google</a> | <a href="/signin/discord">Discord</a>.
         </div>
         <div class="cooldown-timer"></div>
         <div class="palette"></div>
