+++ resources/reference.conf
@@ -72,4 +72,10 @@ oauth {
     key: ""
     secret: ""
   }
+
+  // Create at https://discordapp.com/developers/applications/me
+  discord {
+    key: ""
+    secret: ""
+  }
 }
